appearance of the name in the table

+++ Front/src/components/Admin/HomeArray/index.js
@@ -20,14 +20,14 @@ const HomeArray = () => {
   })
 
   function handleDelete(id) {
-
+  if( window.confirm("Êtes-vous sûr de vouloir supprimer cette ligne ?")){
     axios({
       method: 'delete',
       url: `https://projet-peintre.herokuapp.com/admin/home/${id}`,
     }).then(res => {
       const images = res.data;
       setImages(images);
-    }).catch((err) => { console.log(err) })
+    }).catch((err) => { console.log(err) })}
   }
 
   return (
@@ -66,7 +66,7 @@ const HomeArray = () => {
             <tbody>
               {images.map(image => (
                 <tr>
-                  <td key={image.id_picture}>{image.name_picture}</td>
+                  <td key={image.name_picture}>{image.name_picture}</td>
 
                   <td key={image.image}>{image.image}</td>
